feat(ui): Refactor NodeDetailScreen into smaller components
This commit refactors the monolithic `NodeDetailScreen.kt` into smaller, more manageable, and reusable composable components.

The following components have been extracted:
- `NodeDetailContent`: The main content wrapper.
- `AdministrationSection`: For remote admin and firmware info.
- `DeviceActions`: User actions like favorite, ignore, share, etc.
- `DeviceDetailsContent`: Shows hardware-specific information.
- `EnvironmentMetrics`: Displays environmental sensor data.
- `FirmwareReleaseSheetContent`: Bottom sheet for firmware release notes.
- `InfoCard`: A generic card for displaying metric data.
- `InlineMap`: A small map view for the node's position.
- `MetricsSection`: Groups environment, power, and log sections.
- `NodeDetailsContent`: Core details of the node (name, ID, role).
- `NotesSection`: An input field for user notes on a favorite node.
- `PositionSection`: Displays position, distance, and related actions.
- `PowerMetrics`: Shows power-related metrics.
- `RemoteDeviceActions`: Actions specific to remote nodes.

This refactoring improves code organization, readability, and maintainability of the node detail feature.
