+++ app/src/main/java/com/geeksville/mesh/ui/node/components/DeviceDetailsContent.kt
@@ -1,0 +1,108 @@
+/*
+ * Copyright (c) 2025 Meshtastic LLC
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package com.geeksville.mesh.ui.node.components
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.ColumnScope
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Router
+import androidx.compose.material.icons.twotone.Verified
+import androidx.compose.material3.MaterialTheme.colorScheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.res.vectorResource
+import androidx.compose.ui.unit.dp
+import coil3.compose.AsyncImage
+import coil3.request.ImageRequest
+import com.geeksville.mesh.R
+import com.geeksville.mesh.model.MetricsState
+import org.meshtastic.core.model.DeviceHardware
+import org.meshtastic.core.ui.component.SettingsItemDetail
+import org.meshtastic.core.ui.theme.StatusColors
+
+@Composable
+internal fun ColumnScope.DeviceDetailsContent(state: MetricsState) {
+    val node = state.node ?: return
+    val deviceHardware = state.deviceHardware ?: return
+    val hwModelName = deviceHardware.displayName
+    val isSupported = deviceHardware.activelySupported
+    Box(
+        modifier =
+        Modifier.align(Alignment.CenterHorizontally)
+            .size(100.dp)
+            .clip(CircleShape)
+            .background(color = Color(node.colors.second).copy(alpha = .5f), shape = CircleShape),
+        contentAlignment = Alignment.Center,
+    ) {
+        DeviceHardwareImage(deviceHardware, Modifier.fillMaxSize())
+    }
+
+    Spacer(modifier = Modifier.height(16.dp))
+
+    SettingsItemDetail(
+        text = stringResource(org.meshtastic.core.strings.R.string.hardware),
+        icon = Icons.Default.Router,
+        supportingText = hwModelName,
+    )
+    SettingsItemDetail(
+        text =
+        if (isSupported) {
+            stringResource(org.meshtastic.core.strings.R.string.supported)
+        } else {
+            stringResource(org.meshtastic.core.strings.R.string.supported_by_community)
+        },
+        icon =
+        if (isSupported) {
+            Icons.TwoTone.Verified
+        } else {
+            ImageVector.vectorResource(R.drawable.unverified)
+        },
+        supportingText = null,
+        iconTint = if (isSupported) colorScheme.StatusColors.StatusGreen else colorScheme.StatusColors.StatusRed,
+    )
+}
+
+@Composable
+private fun DeviceHardwareImage(deviceHardware: DeviceHardware, modifier: Modifier = Modifier) {
+    val hwImg = deviceHardware.images?.getOrNull(1) ?: deviceHardware.images?.getOrNull(0) ?: "unknown.svg"
+    val imageUrl = "https://flasher.meshtastic.org/img/devices/$hwImg"
+    AsyncImage(
+        model = ImageRequest.Builder(LocalContext.current).data(imageUrl).build(),
+        contentScale = ContentScale.Inside,
+        contentDescription = deviceHardware.displayName,
+        placeholder = painterResource(R.drawable.hw_unknown),
+        error = painterResource(R.drawable.hw_unknown),
+        fallback = painterResource(R.drawable.hw_unknown),
+        modifier = modifier.padding(16.dp),
+    )
+}
