+++ app/src/main/java/com/geeksville/mesh/ui/node/components/FirmwareReleaseSheetContent.kt
@@ -1,0 +1,95 @@
+/*
+ * Copyright (c) 2025 Meshtastic LLC
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package com.geeksville.mesh.ui.node.components
+
+import android.content.ActivityNotFoundException
+import android.content.Intent
+import android.widget.Toast
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Download
+import androidx.compose.material.icons.filled.Link
+import androidx.compose.material3.Button
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import androidx.core.net.toUri
+import com.mikepenz.markdown.m3.Markdown
+import org.meshtastic.core.database.entity.FirmwareRelease
+import org.meshtastic.core.strings.R
+import timber.log.Timber
+
+@Composable
+internal fun FirmwareReleaseSheetContent(firmwareRelease: FirmwareRelease) {
+    val context = LocalContext.current
+    Column(
+        modifier = Modifier.verticalScroll(rememberScrollState()).padding(16.dp).fillMaxWidth(),
+        verticalArrangement = Arrangement.spacedBy(8.dp),
+    ) {
+        Text(text = firmwareRelease.title, style = MaterialTheme.typography.titleLarge)
+        Text(text = "Version: ${firmwareRelease.id}", style = MaterialTheme.typography.bodyMedium)
+        Markdown(modifier = Modifier.padding(8.dp), content = firmwareRelease.releaseNotes)
+        Row(modifier = Modifier.fillMaxWidth(), horizontalArrangement = Arrangement.spacedBy(8.dp)) {
+            Button(
+                onClick = {
+                    try {
+                        val intent = Intent(Intent.ACTION_VIEW, firmwareRelease.pageUrl.toUri())
+                        context.startActivity(intent)
+                    } catch (e: ActivityNotFoundException) {
+                        Toast.makeText(context, R.string.error_no_app_to_handle_link, Toast.LENGTH_LONG).show()
+                        Timber.e(e)
+                    }
+                },
+                modifier = Modifier.weight(1f),
+            ) {
+                Icon(imageVector = Icons.Default.Link, contentDescription = stringResource(id = R.string.view_release))
+                Spacer(modifier = Modifier.width(8.dp))
+                Text(text = stringResource(id = R.string.view_release))
+            }
+            Button(
+                onClick = {
+                    try {
+                        val intent = Intent(Intent.ACTION_VIEW, firmwareRelease.zipUrl.toUri())
+                        context.startActivity(intent)
+                    } catch (e: ActivityNotFoundException) {
+                        Toast.makeText(context, R.string.error_no_app_to_handle_link, Toast.LENGTH_LONG).show()
+                        Timber.e(e)
+                    }
+                },
+                modifier = Modifier.weight(1f),
+            ) {
+                Icon(imageVector = Icons.Default.Download, contentDescription = stringResource(id = R.string.download))
+                Spacer(modifier = Modifier.width(8.dp))
+                Text(text = stringResource(id = R.string.download))
+            }
+        }
+    }
+}
